Fixed bug #19911 - Select File Dialog causes NullPointer exception.

+++ xml.tools/src/org/netbeans/modules/xml/tools/generator/SelectFileDialog.java
@@ -167,7 +167,8 @@ public class SelectFileDialog extends JPanel {
             fs.runAtomicAction (new FileSystem.AtomicAction () {
                 public void run () throws IOException {
 
-                    if ( tempFile.isVirtual() ) {
+                    if ( ( tempFile != null ) &&
+                         tempFile.isVirtual() ) {
                         tempFile.delete();
                     }
                     folder.createData (newName, ext);
